Move invalid JSON tests to dedicated file
